Restructure move Feature to tools/common package and Application to application package
