+++ src/main/groovy/com/cloudogu/gitops/features/argocd/ArgoCD.groovy
@@ -1,6 +1,5 @@
 package com.cloudogu.gitops.features.argocd
 
-import com.cloudogu.gitops.Feature
 import com.cloudogu.gitops.application.orchestration.GitHandler
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.infrastructure.git.GitRepoFactory
@@ -8,6 +7,7 @@ import com.cloudogu.gitops.infrastructure.helm.HelmClient
 import com.cloudogu.gitops.infrastructure.kubernetes.api.K8sClient
 import com.cloudogu.gitops.infrastructure.kubernetes.rbac.RbacDefinition
 import com.cloudogu.gitops.infrastructure.kubernetes.rbac.Role
+import com.cloudogu.gitops.tools.common.Feature
 import com.cloudogu.gitops.utils.FileSystemUtils
 import com.cloudogu.gitops.utils.MapUtils
