+++ src/test/groovy/com/cloudogu/gitops/cli/GitopsPlaygroundCliTest.groovy
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat
 import static org.mockito.ArgumentMatchers.any
 import static org.mockito.Mockito.*
 
-import com.cloudogu.gitops.Application
+import com.cloudogu.gitops.application.Application
 import com.cloudogu.gitops.config.ApplicationConfigurator
 import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.destroy.Destroyer
