+++ src/test/groovy/com/cloudogu/gitops/config/ApplicationConfiguratorTest.groovy
@@ -1,4 +1,4 @@
-package com.cloudogu.gitops
+package com.cloudogu.gitops.config
 
 import static com.github.stefanbirkner.systemlambda.SystemLambda.withEnvironmentVariable
 import static groovy.test.GroovyAssert.shouldFail
@@ -6,9 +6,6 @@ import static org.assertj.core.api.Assertions.assertThat
 
 import com.cloudogu.gitops.application.content.ContentLoader
 import com.cloudogu.gitops.application.orchestration.GitHandler
-import com.cloudogu.gitops.config.ApplicationConfigurator
-import com.cloudogu.gitops.config.CommonFeatureConfig
-import com.cloudogu.gitops.config.Config
 import com.cloudogu.gitops.config.scm.ScmTenantSchema
 import com.cloudogu.gitops.features.Jenkins
 import com.cloudogu.gitops.features.argocd.ArgoCD
