+++ src/test/groovy/com/cloudogu/gitops/dependencyinjection/okhttp/RetryInterceptorTest.groovy
@@ -1,10 +1,8 @@
-package com.cloudogu.gitops.okhttp
+package com.cloudogu.gitops.dependencyinjection.okhttp
 
 import static com.github.tomakehurst.wiremock.client.WireMock.*
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig
 import static org.assertj.core.api.Assertions.assertThat
-
-import com.cloudogu.gitops.dependencyinjection.okhttp.RetryInterceptor
 
 import javax.net.ssl.HostnameVerifier
 import javax.net.ssl.SSLContext
